requirements.txt
Package             Version
------------------- -------
asgiref             3.2.3  
astroid             2.3.3  
autopep8            1.5    
colorama            0.4.3  
Django              3.0.3  
djangorestframework 3.11.0 
isort               4.3.21 
lazy-object-proxy   1.4.3  
mccabe              0.6.1  
pip                 20.0.2 
pycodestyle         2.5.0  
pylint              2.4.4  
pytz                2019.3 
setuptools          45.2.0 
six                 1.14.0 
sqlparse            0.3.1  
wheel               0.34.2 
wrapt               1.11.2 